requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Flask==1.1.2
Flask-WTF==0.14.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
python-dotenv==0.13.0
requests==2.23.0
six==1.14.0
spotipy==2.11.2
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.3.0
